requirements.txt
streamlit
transformers
torch
faiss-cpu
PyMuPDF
langchain
huggingface_hub